Add Pensjonforvalter integration and enhance configuration
Introduce automated tests for PensjonforvalterHelper. Create PensjonPdlPersonService for fetching extended person data from PDL. Update ApplicationConfig to include a configurable client timeout. Expand RsDollyBestilling to include a check for Pensjon data. Consolidate logic in PensjonforvalterClient by leveraging new services and managing timeouts effectively. Ensure proper handling of related entities and integrate error handling for improved resilience.
