+++ apps/dolly-backend/src/main/java/no/nav/dolly/config/ApplicationConfig.java
@@ -1,15 +1,18 @@
 package no.nav.dolly.config;
 
 import jakarta.annotation.PostConstruct;
+import lombok.Getter;
 import no.nav.testnav.libs.reactivecore.config.CoreConfig;
 import no.nav.testnav.libs.servletcore.config.ApplicationCoreConfig;
 import no.nav.testnav.libs.servletsecurity.config.SecureOAuth2ServerToServerConfiguration;
 import no.nav.testnav.libs.standalone.servletsecurity.config.InsecureJwtServerToServerConfiguration;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.security.core.context.SecurityContextHolder;
 
+@Getter
 @Configuration
 @EnableRetry
 @Import({
@@ -19,6 +22,9 @@ import org.springframework.security.core.context.SecurityContextHolder;
         CoreConfig.class
 })
 public class ApplicationConfig {
+
+    @Value("${dolly.client.general.timeout}")
+    private Long clientTimeout;
 
     @PostConstruct
     public void enableAuthCtxOnSpawnedThreads() {
