+++ apps/dolly-backend/src/main/java/no/nav/dolly/domain/resultset/RsDollyBestilling.java
@@ -132,6 +132,11 @@ public class RsDollyBestilling {
     }
 
     @JsonIgnore
+    public boolean isPensjon() {
+        return nonNull(pensjonforvalter);
+    }
+
+    @JsonIgnore
     public boolean isNonEmpty() {
 
         return Arrays.stream(RsDollyBestilling.class.getMethods())
