+++ apps/dolly-backend/src/main/java/no/nav/dolly/domain/resultset/SystemTyper.java
@@ -17,21 +17,22 @@ public enum SystemTyper {
     ARENA_DAGP("Arena dagpenger"),
     BRREGSTUB("Brønnøysundregistrene (BRREGSTUB)"),
     DOKARKIV("Dokumentarkiv (JOARK)"),
+    FULLMAKT("Fullmakt (Representasjon)"),
     HISTARK("Saksmappearkiv (HISTARK)"),
     INNTK("Inntektskomponenten (INNTK)"),
     INNTKMELD("Inntektsmelding (ALTINN/JOARK)"),
     INST2("Institusjonsopphold (INST2)"),
     KONTOREGISTER("Bankkontoregister"),
     KRRSTUB("Digital kontaktinformasjon (DKIF)"),
-    FULLMAKT("Fullmakt (Representasjon)"),
     MEDL("Medlemskap (MEDL)"),
     ORGANISASJON_FORVALTER("Enhetsregisteret (EREG)"),
     PDLIMPORT("Import av personer (TESTNORGE)"),
     PDL_FORVALTER("Opprett persondetaljer"),
     PDL_ORDRE("Ordre til PDL"),
     PDL_PERSONSTATUS("Person finnes i PDL"),
-    PEN_AP("Alderspensjon (AP)"),
     PEN_AFP_OFFENTLIG("AFP offentlig (PEN)"),
+    PEN_ANNET("Pensjon (PEN)"),
+    PEN_AP("Alderspensjon (AP)"),
     PEN_FORVALTER("Pensjon persondata (PEN)"),
     PEN_INNTEKT("Pensjonsopptjening (POPP)"),
     PEN_PENSJONSAVTALE("Pensjonsavtale (PEN)"),
