+++ apps/dolly-backend/src/main/resources/application-local.yaml
@@ -1,3 +1,8 @@
+dolly:
+  client:
+    general:
+      timeout: 30
+
 spring:
   cache:
     type: none
